requirements.txt
asgiref==3.8.1
certifi==2025.6.15
chardet==3.0.4
charset-normalizer==3.4.2
cwcwidth==0.1.10
Django==5.2.3
django-cors-headers==4.7.0
djangorestframework==3.16.0
googletrans==4.0.0rc1
gunicorn==23.0.0
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2025.1.1
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
lxml==6.0.0
packaging==25.0
pillow==11.2.1
psycopg2-binary==2.9.10
reportlab==4.4.2
requests==2.32.4
rfc3986==1.5.0
sniffio==1.3.1
sqlparse==0.5.3
typing_extensions==4.14.0
urllib3==2.5.0
whitenoise==6.9.0
